調整連接類型的屬性為浮點數，更新速度、延遲、可靠性和距離的值；修改 CityStreetSceneManager 中的 SpeedUp 方法參數類型，並將 Client 中的錯誤日誌更改為警告

+++ Assets/Scripts/CityStreetSceneManager.cs
@@ -158,7 +158,7 @@ namespace InternetEmpire
             );
         }
 
-        public void SpeedUp(int speed)
+        public void SpeedUp(float speed)
         {
             Time.timeScale = speed; // 將遊戲速度加倍
         }
